+++ utbot-instrumentation/src/main/kotlin/org/utbot/instrumentation/process/InstrumentedProcessMain.kt
@@ -6,19 +6,14 @@ import com.jetbrains.rd.util.reactive.adviseOnce
 import kotlinx.coroutines.*
 import org.mockito.Mockito
 import org.utbot.common.*
-import org.utbot.framework.plugin.api.ClassId
 import org.utbot.framework.plugin.api.util.UtContext
 import org.utbot.framework.process.kryo.KryoHelper
 import org.utbot.instrumentation.agent.Agent
 import org.utbot.instrumentation.instrumentation.Instrumentation
 import org.utbot.instrumentation.instrumentation.coverage.CoverageInstrumentation
-import org.utbot.instrumentation.instrumentation.spring.SpringUtExecutionInstrumentation
 import org.utbot.instrumentation.process.generated.CollectCoverageResult
-import org.utbot.instrumentation.process.generated.GetSpringBeanResult
-import org.utbot.instrumentation.process.generated.GetSpringRepositoriesResult
 import org.utbot.instrumentation.process.generated.InstrumentedProcessModel
 import org.utbot.instrumentation.process.generated.InvokeMethodCommandResult
-import org.utbot.instrumentation.process.generated.TryLoadingSpringContextResult
 import org.utbot.instrumentation.process.generated.instrumentedProcessModel
 import org.utbot.rd.IdleWatchdog
 import org.utbot.rd.ClientProtocolBuilder
@@ -149,6 +144,7 @@ private fun InstrumentedProcessModel.setup(kryoHelper: KryoHelper, watchdog: Idl
         Agent.dynamicClassTransformer.useBytecodeTransformation = params.useBytecodeTransformation
         Agent.dynamicClassTransformer.transformer = instrumentation
         Agent.dynamicClassTransformer.addUserPaths(pathsToUserClasses)
+        instrumentation.run { setupAdditionalRdResponses(kryoHelper, watchdog) }
     }
     watchdog.measureTimeForActiveCall(addPaths, "User and dependency classpath setup") { params ->
         pathsToUserClasses = params.pathsToUserClasses.split(File.pathSeparatorChar).toSet()
@@ -161,19 +157,5 @@ private fun InstrumentedProcessModel.setup(kryoHelper: KryoHelper, watchdog: Idl
         logger.debug { "class - ${anyClass.name}" }
         val result = (instrumentation as CoverageInstrumentation).collectCoverageInfo(anyClass)
         CollectCoverageResult(kryoHelper.writeObject(result))
-    }
-    watchdog.measureTimeForActiveCall(getSpringBean, "Getting Spring bean") { params ->
-        val springUtExecutionInstrumentation = instrumentation as SpringUtExecutionInstrumentation
-        val beanModel = springUtExecutionInstrumentation.getBeanModel(params.beanName, pathsToUserClasses)
-        GetSpringBeanResult(kryoHelper.writeObject(beanModel))
-    }
-    watchdog.measureTimeForActiveCall(getRelevantSpringRepositories, "Getting Spring repositories") { params ->
-        val classId: ClassId = kryoHelper.readObject(params.classId)
-        val repositoryDescriptions = (instrumentation as SpringUtExecutionInstrumentation).getRepositoryDescriptions(classId)
-        GetSpringRepositoriesResult(kryoHelper.writeObject(repositoryDescriptions))
-    }
-    watchdog.measureTimeForActiveCall(tryLoadingSpringContext, "Trying to load Spring application context") { params ->
-        val contextLoadingResult = (instrumentation as SpringUtExecutionInstrumentation).tryLoadingSpringContext()
-        TryLoadingSpringContextResult(kryoHelper.writeObject(contextLoadingResult))
     }
 }
